finished statistics min-max extreme, included as -1000, +1000 in percentile function
