+++ api/time_series.cpp
@@ -1,5 +1,6 @@
 #include "core/core_pch.h"
 #include "time_series.h"
+
 #include <dlib/statistics.h>
 
 namespace shyft{
