+++ api/time_series.h
@@ -4,7 +4,7 @@
 #include "core/utctime_utilities.h"
 #include "core/time_axis.h"
 #include "core/time_series.h"
-
+#include "core/time_series_statistics.h"
 namespace shyft {
     namespace api {
         using namespace shyft::core;
