+++ core/core.vcxproj.filters
@@ -129,6 +129,9 @@
     <ClInclude Include="time_series.h">
       <Filter>time_series</Filter>
     </ClInclude>
+    <ClInclude Include="time_series_statistics.h">
+      <Filter>time_series</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="methods">
